requirements.txt
Pillow
stego-lsb